requirements.txt
ST7789
Pillow
RPi.GPIO
spidev
numpy
gpiozero
lgpio
